+++ CHANGELOG.md
@@ -6,6 +6,8 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 ## [Unreleased]
 
+- Imagick reference updated to version 3.4.1 (stable)
+
 ## [1.6.0] - 2016-03-05
 
 ### Added